feat(expenses): enforce integer product quantities
  Store expense item quantities as integers, reject decimal quantity payloads, and migrate existing decimal quantities by rounding them up.

  Update expenses and insights tests to match the integer quantity contract

+++ internal/insights/repository_test.go
@@ -55,7 +55,7 @@ func newTestInsightsRepository(t *testing.T) Repository {
 		CREATE TABLE expense_items (
 			id INTEGER PRIMARY KEY AUTOINCREMENT,
 			expense_id INTEGER NOT NULL,
-			quantity REAL NOT NULL,
+			quantity INTEGER NOT NULL,
 			unit_price REAL NOT NULL
 		);
